+++ layers/CMSSchema/packages/settings/src/TypeAPIs/SettingsTypeAPIInterface.php
@@ -13,6 +13,12 @@ interface SettingsTypeAPIInterface
      * @throws OptionNotAllowedException When the option does not exist, or is not in the allowlist
      */
     public function getOption(string $name, array $options = []): mixed;
+    /**
+     * List of all the option names stored in the DB.
+     *
+     * @return string[]
+     */
+    public function getOptionNames(): array;
     public function validateIsOptionAllowed(string $key): bool;
     /**
      * @return string[]